Move spot_instance_option_manager under minion_manager.
