+++ platform/source/lib/ax/platform/minion_manager/__init__.py
@@ -1,0 +1,1 @@
+from .spot_instance_option_manager import SpotInstanceOptionManager
